flash, nor, nrfx: add support for both hwid based and part based devices
Some nordic devices (nRF51) implement FICR.CONFIGID, while others (nRF52, nRF91)
have FICR.PART, FICR.VARIANT, ..... Add code for supporting both of them.

Looks like the original nrf5 driver used CONFIGID for the nRF52 too, but
that seems deprecated (see https://devzone.nordicsemi.com/f/nordic-q-a/17105/is-there-a-possibility-to-read-out-the-hwid-and-map-it-to-ic-revision-and-build-code-nrf52) for more details.